# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stager API
  description: |
    API for Stager - AI-Powered Virtual Staging Platform for Real Estate.

    ## Authentication
    All endpoints (except webhooks) require authentication via Supabase Auth session cookies.
    Unauthenticated requests return `401 Unauthorized`.

    ## Rate Limiting
    Resource-intensive endpoints are rate-limited:
    - **Staging:** 20 requests/minute
    - **Email:** 10 requests/minute
    - **Auth:** 5 requests/minute

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Error Responses
    All errors follow a consistent format with actionable suggestions:
    ```json
    {
      "error": "Human-readable message",
      "code": "ERROR_CODE",
      "suggestion": "Actionable suggestion for the user",
      "action": { "label": "Button label", "href": "/path" },
      "details": { "field": "value" }
    }
    ```

    ## Credits
    - Staging costs 1 credit per image
    - First 2 remixes per original image are free
    - Additional remixes cost 1 credit each
  version: 1.136.0
  contact:
    name: Stager Support
servers:
  - url: https://stager-tau.vercel.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Local Development
...
paths:
...
  /team/invite:
    post:
      tags: [Team]
      summary: Invite team member
      description: |
        Send an email invitation to join the organization.
        Only organization owners can invite members.

        **Rate Limit:** 10 emails per minute.
      operationId: inviteTeamMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamInviteRequest'
      responses:
        '200':
          description: Invitation sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  invitation:
                    $ref: '#/components/schemas/Invitation'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Invitation already exists or team full
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /team/members/{id}:
    delete:
      tags: [Team]
      summary: Remove team member
      description: Remove a member from the organization. Only organization owners can remove members.
      operationId: removeTeamMember
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Member ID (organization_members.id, not user ID)
      responses:
        '200':
          description: Member removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /team/members/{id}/credits:
    patch:
      tags: [Team]
      summary: Allocate credits to member
      description: |
        Update the credit allocation for a team member.
        Credits are transferred from the organization's unallocated pool.
      operationId: allocateMemberCredits
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credits]
              properties:
                credits:
                  type: integer
                  minimum: 0
                  description: New credit allocation for the member
      responses:
        '200':
          description: Credits allocated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  newBalance:
                    type: integer
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ============ SEARCH ============
...
components:
...
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - UNAUTHORIZED
        - SESSION_EXPIRED
        - FORBIDDEN
        - NOT_OWNER
        - INSUFFICIENT_PERMISSIONS
        - INSUFFICIENT_CREDITS
        - PAYMENT_REQUIRED
        - RATE_LIMITED
        - VALIDATION_ERROR
        - INVALID_INPUT
        - NOT_FOUND
        - ALREADY_EXISTS
        - CONFLICT
        - PROVIDER_ERROR
        - EMAIL_ERROR
        - INTERNAL_ERROR
        - SERVICE_UNAVAILABLE
      description: Standardized error codes

    # ---- Request Schemas ----
...
    TeamInviteRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Email address to invite
        initialCredits:
          type: integer
          minimum: 0
          default: 0
          description: Credits to allocate from organization pool

    # ---- Response Schemas ----
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          $ref: '#/components/schemas/ErrorCode'
        suggestion:
          type: string
          description: Actionable suggestion for the user
        action:
          type: object
          description: Optional CTA button
          properties:
            label:
              type: string
            href:
              type: string
        details:
          type: object
          additionalProperties: true
          description: Additional context about the error
...
    Invitation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        initialCredits:
          type: integer
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "You must be logged in to perform this action."
            code: "UNAUTHORIZED"
            suggestion: "Please log in or create an account to continue."
            action:
              label: "Log In"
              href: "/login"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "This action is restricted to organization owners."
            code: "INSUFFICIENT_PERMISSIONS"
            suggestion: "Contact your organization owner for assistance."
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "The requested resource could not be found."
            code: "NOT_FOUND"
    ValidationError:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Please fix the errors in your submission."
            code: "VALIDATION_ERROR"
            suggestion: "Review the highlighted fields and correct any issues."
            details:
              fields:
                - field: "email"
                  message: "Invalid email format"
...
    RateLimited:
      description: Too many requests (HTTP 429)
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Maximum requests per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Too many requests. Please slow down."
            code: "RATE_LIMITED"
            suggestion: "Try again in 1 minute."
            details:
              resetInSeconds: 60
...